fix(savestates): fix unfinished dma being present when loading an old mupen st while st task queue contains save followed by load on pure interp
When loading a savestate on pure interpreter with the "Undo Savestate Load" option, which prepends a save operation to the st task queue, the DMA finishing code wouldn't run due to the cart hack bit missing. This fixes the situation by circumventing the need for a hack bit check by employing a more lenient and potentially crappy unified check for an SI interrupt which determines whether the DMA finishing is performed.