test(Benchmark): benchmark functionality (#301)
# Summary

Adds benchmarking functionality to Mupen using the `--benchmark {PATH}`
or `-b {PATH}` commandline parameter.

Also adds a ci step which runs benchmarks in the cloud automatically and
displays information about detected performance improvements or
regressions. The script responsible for this functionality can also be
run locally.

Mupen currently crashes in the cloud runner, but that's a separate bug.