fix(AVIEncoder): AVI file stream not being closed when cancelling the codec dialog
Closes #309.

Also changes the return value of `Encoder::start` to allow failing with no error message, so that cancelling the codec dialog stops the encode starting process without showing a misleading error dialog.

Also slightly refactors AVIEncoder internally to delete the output file when stopping due to a fail or cancellation in order to avoid an output file being created when the user doesn't expect it.