add hotkey locking functionality to ActionManager
rationale:

When a Lua script enters a state that makes it inappropriate to trigger hotkeys in (e.g. typing in a textbox), it can't choose to temporarily block hotkeys from triggering that would otherwise conflict with the current operation.

summary:

adds `ActionManager::lock_hotkeys` and `ActionManager::get_hotkeys_locked`

exposed to Lua with test coverage